requirements.txt
python-telegram-bot[job-queue]==20.7
python-telegram-bot-raw==20.7
flask==2.0.3
aiosqlite==0.19.0
python-dotenv==1.0.0